Update repartition functions to allow for Col in numPartitions parameter.
Reference: https://spark.apache.org/docs/latest/api/python/pyspark.sql.html#pyspark.sql.DataFrame.repartition

numPartitions – can be an int to specify the target number of partitions or a Column.
If it is a Column, it will be used as the first partitioning column.
If not specified, the default number of partitions is used.